% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{print.close_reply}
\alias{print.close_reply}
\title{Print a close_reply}
\usage{
\method{print}{close_reply}(x, ...)
}
\arguments{
\item{x}{A \link{close_reply}.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a close_reply
}
\keyword{internal}
